Separate HTTP preview code and URL previewer.
